Moves query migration details to separate file
